Only show KPIs where goals are set
This is a temporary change so that we can show the current state
of the KPIs in staging/production. It's an awkward solution and the
intention is to rework the template to dynamically generate each
table using snippets (and also to only call the stats functions
when goals have been set).